change amount type from int to float64

+++ Application/v1/viewModels/CurrencyViewModel.go
@@ -7,9 +7,9 @@ import (
 )
 
 type JsonCurrency struct {
-	Amount          int    `json:"amount"`
-	CurrentCurrency string `json:"currentCurrency"`
-	WantedCurrency  string `json:"wantedCurrency"`
+	Amount          float64 `json:"amount"`
+	CurrentCurrency string  `json:"currentCurrency"`
+	WantedCurrency  string  `json:"wantedCurrency"`
 }
 
 type CurrencyViewModel struct {
@@ -17,7 +17,7 @@ type CurrencyViewModel struct {
 	validation validation.Validation
 }
 
-func (c *CurrencyViewModel) GetAmount() int {
+func (c *CurrencyViewModel) GetAmount() float64 {
 	return c.Amount
 }
